Move Francken, App and Http folder into src folder
The reasoning behind this commit is to have less cluttering in the root
folder. In a next commit we will change some of the namespacing, i.e.
use a "Francken" root namespace and a "Francken\Domain",
"Francken\Application" and "Francken\Infrastructure" namespaces.

Changing the namespaces and restructuring some of the files so that we
get a nice hexagonal architecture requires a bit more work so that will
be changed in a later commit.