Refactor App and Http namespaces
We now use the Infrastructure namespace for objects that need knowledge
of external services, for instance databases, http requests or service
containers.

The application namespace contains application logic such as handling a
command or a published event.
This namespace may also contain processors such as notifiers (using for
instance an Email service from the infrastructure namespace).