Make providing PaymentInfo optional
When someone signs up to be a member he or she will have to provide a
signature and agree that a membership fee will be withdrawn from their
bank account.
Since this can be done after submitting the registration request we make
this optional.
Note that when accepting the registration request the payment info
should have been provided.