Encapsulate database access through specific repositories
Though the result of this commit is additional boilerplate for read models we no
longer have to specify a specific read model repository implementation for any
class that would be needing a ReadModelRepository.
Instead we now ask for a specific repository which is configured inside of one
of our service providers.