Instead of showing title and description show a card with a title
This commit partially implements issue #104.
The overview of the committees are now shown in a grid of cards that either show
the title of a committee, or their logo.

Currently only the logo of the borrelcie is provided.
Later we should be able to manage committees from the admin page and upload
their logo there.