Move static committee pages to dynamic committee pages
Committee pages are now handled by the CommitteesController which currently
contains a static inmemory array containing the committees.
This will be changed so that the data is stored in a repository.

Each committee page shows a list of its members and an aside element containing
a list of all committees.

The content of a committee page is determined by its static page that was used
earlier.
Later we might want to change this so that a committee can either have a page
using compiled markdown, or a custom page so that the committee can have custom
css and javascript (such as the Brouwcie).