Use an inmemory BookDetailsRepository for offline applications
The books specific code in the AppServiceProvider has now been placed in
BookServiceProvider, which will use the InMemoryBookDetailsRepository when
the environment variable APP_OFFLINE is set to true.
This was made because the previous provider can only retrieve details once in a
couple of seconds / minutes due to Amazon's throtling of web requests.
We should probably just use a real API for retrieving the details.
The naming of the repositories has been made somewhat more consistent.