Extract footer sponsors as a blade component
This nice thing about this is that we don't need to use a view composer
to get the footer logos.

Instead the component allows us to be more explicit about the data
needed for its view.

Later we might also add caching here when getting the logos from a
database, though we should be careful not to perform too many side
effects here.