Configure Nextcloud as a filesystem
In our Nextcloud implementation we will use flysystem and their
WebDAVAdapter to make the remote Nextcloud instance act as a local
filesystem.

This will allow us to easily fetch directories and files from nextcloud
as if they are normal files from our local filesystem.
Not only that it also makes writing unit/integration tests quite simple
as Laravel has some nice tools to fake a storage implementation.

A downside of not using their official api might be that we won't be
able to fetch more specific data, i.e. reuse the album capabilities that
Nextcloud has.
This seems to be fine for now though, maybe this is something we can
look into at another time.