Removed board room number and added BBD
I removed the phone number of the board room from the footer and the contact page. Just like the external phone number, the phone number of the chair can be entered in admin and it replaces the number on the footer and contact page. Also BBD can be set in the header.